# Optimize streaming events receiving