Небольшая оптимизация получения стриминг эвентов.
Результаты синтетического теста. На ~30% меньше выделение памяти и на ~7% меньше расхода CPU.

|    Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|  Original | 8.796 μs | 0.0766 μs | 0.0679 μs |  1.00 | 0.9918 |     - |     - |   3.08 KB |
| Optimized | 8.199 μs | 0.1015 μs | 0.0900 μs |  0.93 | 0.7019 |     - |     - |   2.17 KB |