@olsh 2021-01-12
@rus-art я обновил PR. Использовал подход где внутренний массив в MemoryStream не резайзится (почти). Думаю что это будет лучше подхода с `SetLength` в моем первоначальном варианте и вашем.
Дайте знать что думаете по этому поводу.

Финальные замеры

|    Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|  Original | 8.744 μs | 0.0476 μs | 0.0422 μs |  1.00 | 0.9918 |     - |     - |   3.08 KB |
| Optimized | 7.796 μs | 0.0445 μs | 0.0371 μs |  0.89 | 0.6104 |     - |     - |   1.88 KB |

Вторая серьезная аллокация происходит [тут](https://github.com/TinkoffCreditSystems/invest-openapi-csharp-sdk/blob/fc134a644931574bb3499864008963dcde99ca35/src/Tinkoff.Trading.OpenApi/Models/StreamingResponseConverter.cs#L44), предлагаю исправить ее отдельным PR.