@rus-art 2021-01-18
`SetLength` не делает ресайз массива, если новая длина не больше старой, так что в его использовании в нашем случае нет ничего плохого