@olsh 2021-02-19
@rus-art я обновил PR, теперь копирование не будет происходить если сообщение полностью поместилось в буффер (на самом деле я не видел сообщений больше буффера, так что, вероятно, всегда будет так). MemoryStream используется как фоллбек если все таки пришло большое сообщение.